[mc_log_ui] Minor changes for Python 3
Didn't test everything but should be working for the most part

+++ utils/mc_log_gui/mc_log_plotcanvas.py
@@ -73,7 +73,7 @@ class PlotFigure(object):
     # self.colors = ['r', 'g', 'b', 'y', 'k', 'cyan', 'magenta', 'orange']
     cm = matplotlib.cm.Set1
     cm2rgb = (np.array(cm(x)[0:3]) for x in np.linspace(0, 1, self.Ncolor))
-    self.colors = ['#%02x%02x%02x' % tuple(255 * rgb) for rgb in cm2rgb]
+    self.colors = ['#%02x%02x%02x' % tuple((255 * rgb).astype(int)) for rgb in cm2rgb]
 
     self.x_data = 't'
